今日のチェックイン済みドッグランの取得処理作成.clain内roleをintに変更(#101)

+++ internal/auth/core/dto/auth_user_info_dto.go
@@ -3,5 +3,5 @@ package dto
 type UserAuthInfoDTO struct {
 	UserID int64
 	JwtID  string
-	RoleID int64
+	RoleID int
 }
